microservicio/infraestructura/src/test/java/com/ceiba/agenda/controlador/ConsultaControladorAgendaTest.java
package com.ceiba.agenda.controlador;

import com.ceiba.ApplicationMock;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.http.MediaType;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static org.hamcrest.Matchers.hasSize;
import static org.hamcrest.core.Is.is;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(SpringRunner.class)
@ContextConfiguration(classes= ApplicationMock.class)
@WebMvcTest(ConsultaControladorAgenda.class)
public class ConsultaControladorAgendaTest {
    @Autowired
    private MockMvc mocMvc;


    @Test
    public void listar() throws Exception {
        // arrange

        // act - assert
        mocMvc.perform(get("/agendas")
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$", hasSize(1)))
                .andExpect(jsonPath("$[0].idUsuario", is(1)));
    }

    @Test
    public void listarPorUsuario() throws Exception {
        // arrange
        Long id =1L;
        // act - assert
        mocMvc.perform(get("/agendas/usuario/{id}",id)
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$", hasSize(1)))
                .andExpect(jsonPath("$[0].idUsuario", is(1)));
    }

    @Test
    public void buscarPorId() throws Exception {
        // arrange
        Long id = 1L;
        // act - assert
        mocMvc.perform(get("/agendas/{id}",id)
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("id", is(1)));
    }
}
